Moved Redux Stuff Into Folder
Part of #32
