+++ src/redux/selectors/nowPlaying.ts
